+++ examples/complete/azure/autoscale/main.tf
@@ -47,6 +47,18 @@ data "hopsworksai_instance_type" "small_worker" {
   min_cpus       = 4
 }
 
+resource "azurerm_container_registry" "acr" {
+  name                = "tfhopsworksbasic"
+  resource_group_name = module.azure.resource_group
+  location            = module.azure.location
+  sku                 = "Premium"
+  admin_enabled       = false
+  retention_policy {
+    enabled = true
+    days    = 7
+  }
+}
+
 resource "hopsworksai_cluster" "cluster" {
   name    = "tf-hopsworks-cluster"
   ssh_key = module.azure.ssh_key_pair_name
@@ -73,6 +85,7 @@ resource "hopsworksai_cluster" "cluster" {
     container {
       storage_account = module.azure.storage_account_name
     }
+    acr_registry_name = azurerm_container_registry.acr.name
   }
 
   rondb {