+++ test-fixtures/aws/main.tf
@@ -1,5 +1,4 @@
 data "hopsworksai_aws_instance_profile_policy" "policy" {
-  enable_eks_and_ecr = false
 }
 
 resource "aws_iam_role" "role" {