+++ CHANGELOG.md
@@ -7,6 +7,7 @@ BREAKING CHANGES:
 ENHANCEMENTS:
 * datasource/aws_instance_profile_policy: Deprecate `enable_eks_and_ecr` attribute to use `enable_eks` and `enable_ecr` instead
 * datasource/azure_user_assigned_identity_permissions: Deprecate `enable_aks_and_acr` attribute  to use `enable_aks` and `enable_acr` instead
+* Update acceptance tests and examples for new release
 
 BUG FIXES: